Use own jsonld script tag for WebSite and DataCatalog
